Fix amount msvc warnings

+++ csdb/src/amount_commission.cpp
@@ -14,14 +14,14 @@ AmountCommission::AmountCommission(double value) {
     u_.fIEEE.sign = value < 0. ? 1 : 0;
     value = std::fabs(value);
     double expf = value == 0. ? 0. : std::log10(value);
-    int expi = expf >= 0. ? expf + 0.5 : expf - 0.5;
+    int expi = static_cast<int>(expf >= 0. ? expf + 0.5 : expf - 0.5);
     value /= std::pow(10, expi);
     if (value >= 1.) {
         value *= 0.1;
         ++expi;
     }
-    u_.fIEEE.exp = expi + 18;
-    u_.fIEEE.frac = lround(value * 1024);
+    u_.fIEEE.exp = static_cast<uint16_t>(expi + 18);
+    u_.fIEEE.frac = static_cast<uint16_t>(lround(value * 1024));
 }
 
 void AmountCommission::put(priv::obstream& os) const {
